成長管理機能の実装

+++ db/migrate/20220609063012_create_growths.rb
@@ -3,8 +3,8 @@ class CreateGrowths < ActiveRecord::Migration[6.0]
     create_table :growths do |t|
       t.string :month,          null: false   
       t.date :inspection_date,  null: false   
-      t.integer :weight,        null: false 
-      t.integer :height,        null: false 
+      t.float	 :weight,         null: false 
+      t.float	 :height,         null: false 
       t.references :kid,        null: false   
 
       t.timestamps
